requirements.txt
# external requirements
numpy
pandas
torch
pyro-ppl
ipykernel
seaborn
Pillow
kaggle

# local package
-e .

